LPS-145491 Rename assignment data retrieve and move it from DiagramBuilder
