+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-designer-web/src/main/resources/META-INF/resources/designer/js/definition-builder/diagram-builder/util/populateAssignmentData.js
@@ -1,0 +1,63 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ */
+
+import {retrieveRolesBy, retrieveUsersBy} from '../../util/fetchUtil';
+import {getAssignmentType} from '../components/sidebar/sections/assignments/utils';
+
+const populateAssignmentsData = (initialElements, setElements) => {
+	for (let index = 0; index < initialElements.length; index++) {
+		const element = initialElements[index];
+		if (element.data.assignments) {
+			const assignmentType = getAssignmentType(element.data.assignments);
+
+			if (assignmentType === 'user') {
+				const usersData = [];
+
+				retrieveUsersBy(
+					'emailAddress',
+					element.data.assignments.emailAddress
+				)
+					.then((response) => response.json())
+					.then(({items}) => {
+						items.forEach((item, index) => {
+							usersData.push({
+								emailAddress: item.emailAddress,
+								identifier: `${Date.now()}-${index}`,
+								name: item.name,
+								screenName: item.alternateName,
+								userId: item.id,
+							});
+						});
+					})
+					.then(() => {
+						initialElements[
+							index
+						].data.assignments.usersData = usersData;
+						setElements([...initialElements]);
+					});
+			}
+			else if (assignmentType === 'roleId') {
+				retrieveRolesBy('roleId', element.data.assignments.roleId)
+					.then((response) => response.json())
+					.then((response) => {
+						initialElements[index].data.assignments.rolesData = {
+							id: response.id,
+							name: response.name,
+							roleType: response.roleType,
+						};
+						setElements([...initialElements]);
+					});
+			}
+		}
+	}
+};
+
+export default populateAssignmentsData;
